# docs/ui/control-plane-mock/openapi.yaml
openapi: 3.1.0
info:
  title: Quant Research Platform Control Plane UI Mock
  version: 0.2.0
  description: >
    G0 frontend contract baseline. Endpoints marked frontend-proposal require
    backend review before implementation.
  license:
    name: Proprietary
servers:
  - url: http://localhost:4010
    description: Local mock
security:
  - oidc: []
  - bearerAuth: []
...
paths:
...
  /v1/experiments:preregister:
    post:
      tags: [Experiments]
      operationId: preregisterExperiment
      summary: Preregister an experiment specification
      x-source: technical-design
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - metadata
                - job_id
                - candidate_ids
                - validation_policy_version
              properties:
                metadata:
                  $ref: "#/components/schemas/CommandMetadata"
                job_id:
                  type: string
                candidate_ids:
                  type: array
                  minItems: 1
                  items:
                    type: string
                validation_policy_version:
                  type: string
      responses:
        "202":
          $ref: "#/components/responses/CommandAccepted"
  /v1/experiments/{experiment_id}:run:
    post:
      tags: [Experiments]
      operationId: runExperiment
      summary: Run a preregistered experiment
      x-source: technical-design
      parameters:
        - $ref: "#/components/parameters/ExperimentId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommandMetadata"
      responses:
        "202":
          $ref: "#/components/responses/CommandAccepted"
  /v1/experiments/{experiment_id}:
    get:
      tags: [Experiments]
      operationId: getExperiment
      summary: Get an experiment and its runs
      x-source: frontend-proposal
      parameters:
        - $ref: "#/components/parameters/ExperimentId"
      responses:
        "200":
          description: Experiment, current attempt, budget, and heartbeat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Strong ETag of the aggregate version being changed.
      schema:
        type: string
        minLength: 3
...
    ExperimentId:
      name: experiment_id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://identity.example.invalid/.well-known/openid-configuration
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    CommandAccepted:
      description: Command accepted for asynchronous processing
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CommandReceipt"
...
  schemas:
...
    ExperimentSpecState:
      type: string
      enum: [DRAFT, PREREGISTERED, SUPERSEDED, CLOSED]
    ExperimentRunState:
      type: string
      enum:
        - QUEUED
        - RUNNING
        - WAITING_INPUT
        - BLOCKED_POLICY
        - SUCCEEDED
        - FAILED_RETRYABLE
        - FAILED_TERMINAL
        - QUARANTINED
        - NON_REPRODUCIBLE
        - CANCELLED
    AttemptState:
      type: string
      enum: [QUEUED, RUNNING, SUCCEEDED, FAILED, CANCELLED, TIMED_OUT]
...
    Budget:
      type: object
      additionalProperties: false
      required: [candidate_limit, wall_clock_minutes]
      properties:
        candidate_limit:
          type: integer
          minimum: 1
        llm_token_limit:
          type: integer
          minimum: 0
        cpu_hours:
          type: number
          minimum: 0
        wall_clock_minutes:
          type: integer
          minimum: 1
    CommandMetadata:
      type: object
      additionalProperties: false
      required: [reason, parent_artifact_id, budget, schema_version]
      properties:
        reason:
          type: string
          minLength: 3
        parent_artifact_id:
          type: [string, "null"]
        budget:
          $ref: "#/components/schemas/Budget"
        schema_version:
          type: string
...
    ResourceFreshness:
      type: object
      required: [as_of, is_stale]
      properties:
        as_of:
          type: string
          format: date-time
        is_stale:
          type: boolean
        stale_reason:
          type: [string, "null"]
...
    Attempt:
      type: object
      required: [attempt, state]
      properties:
        id:
          type: string
        attempt:
          type: integer
          minimum: 1
        state:
          $ref: "#/components/schemas/AttemptState"
        started_at:
          type: [string, "null"]
          format: date-time
        heartbeat_at:
          type: [string, "null"]
          format: date-time
        error_code:
          type: [string, "null"]
...
    Experiment:
      type: object
      required:
        - id
        - job_id
        - spec_state
        - runs
        - freshness
        - allowed_actions
      properties:
        id:
          type: string
        job_id:
          type: string
        spec_state:
          $ref: "#/components/schemas/ExperimentSpecState"
        runs:
          type: array
          items:
            $ref: "#/components/schemas/ExperimentRun"
        budget:
          $ref: "#/components/schemas/Budget"
        budget_used:
          type: object
          additionalProperties:
            type: number
        freshness:
          $ref: "#/components/schemas/ResourceFreshness"
        allowed_actions:
          type: array
          items:
            type: string
    ExperimentRun:
      type: object
      required:
        - id
        - state
        - run_class
        - attempts
        - dataset_snapshot_id
        - rule_set_snapshot_id
        - run_fingerprint
      properties:
        id:
          type: string
        state:
          $ref: "#/components/schemas/ExperimentRunState"
        run_class:
          type: string
          enum: [EXPLORATORY, FORMAL]
        attempts:
          type: array
          items:
            $ref: "#/components/schemas/Attempt"
        dataset_snapshot_id:
          type: string
        rule_set_snapshot_id:
          type: string
        run_fingerprint:
          type: string
...
    CommandReceipt:
      type: object
      required: [command_id, status, resource_id, submitted_at]
      properties:
        command_id:
          type: string
        status:
          const: ACCEPTED
        resource_id:
          type: string
        submitted_at:
          type: string
          format: date-time
...